+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] - 2025-10-10
+
+## Added
+
+- Added support for case-sensitive and regex searches in [app.rs](src/ui/app.rs) and
+[main.rs](src/main.rs)
+
+## Changed
+
+- Regexes are now cached to improve performance in [app.rs](src/ui/app.rs)
+
 ## [0.6.3] - 2025-09-28
 
 ## Added